fmt: Small cleanup of the landing page sketch
